+++ x/interchainstaking/keeper/abci.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/cosmos/cosmos-sdk/telemetry"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	tmtypes "github.com/cosmos/ibc-go/v5/modules/light-clients/07-tendermint/types"
-
 	"github.com/ingenuity-build/quicksilver/x/interchainstaking/types"
 )
 
@@ -16,7 +16,28 @@ type zoneItrFn func(index int64, zoneInfo types.Zone) (stop bool)
 // BeginBlocker of interchainstaking module
 func (k Keeper) BeginBlocker(ctx sdk.Context) {
 	defer telemetry.ModuleMeasureSince(types.ModuleName, time.Now(), telemetry.MetricKeyBeginBlocker)
+	// post upgrade-v1.2.5 processing
+	if ctx.BlockHeight() == 14540740 {
+		zone, found := k.GetZone(ctx, "regen-1")
+		if found {
+			k.IterateReceipts(ctx, func(_ int64, receiptInfo types.Receipt) (stop bool) {
+				if receiptInfo.ChainId == "regen-1" && receiptInfo.Completed == nil {
+					sendMsg := banktypes.MsgSend{
+						FromAddress: "",
+						ToAddress:   "",
+						Amount:      receiptInfo.Amount,
+					}
+					err := k.handleSendToDelegate(ctx, &zone, &sendMsg, receiptInfo.Txhash)
+					if err != nil {
+						k.Logger(ctx).Error("error in processing Pending delegations for regen-1 ", "error", err)
+					}
 
+				}
+				return false
+			})
+		}
+
+	}
 	if ctx.BlockHeight()%30 == 0 {
 		if err := k.GCCompletedRedelegations(ctx); err != nil {
 			k.Logger(ctx).Error("error in GCCompletedRedelegations", "error", err)