+++ x/interchainstaking/types/codec.go
@@ -29,6 +29,8 @@ func RegisterInterfaces(registry types.InterfaceRegistry) {
 		(*sdk.Msg)(nil),
 		&MsgSignalIntent{},
 		&MsgRequestRedemption{},
+		&MsgGovCloseChannel{},
+		&MsgGovReopenChannel{},
 	)
 
 	registry.RegisterImplementations(